Added functionality to allow the player to shoot
The player can shoot 1 bullet at a time. The bullet will travel up,
in a straight line, until it either hits and enemy or goes off the screen.
If the bullet travels off the screen, the bullet will destroy itself and the player
will be able to shoot again. If the bullet hits an enemy, both the bullet and
enemy will be destroyed.

(have not implemented scoring yet)